requirements.txt
# Should use latest 1.9 for security
Django
pinax-theme-bootstrap
django-user-accounts
metron
pinax-eventlog
django-formset-js
dj-static
dj-database-url
#pinax-pages
pinax-boxes
wagtail
pylibmc
raven
bleach
html5lib
django-debug-toolbar
django-markdownify
#progressbar
lxml

# database
psycopg2

#Wiki
wiki

# For testing
django-nose
coverage

# Registrasion + Symposion
# Install via pip repo cloned locally
#registrasion

-e git+https://github.com/pyconau2017/symposion.git@master#egg=symposion
-e git+https://github.com/pyconau2017/registrasion.git@master#egg=registrasion
-e git+https://github.com/pyconau2017/registrasion-stripe.git@master#egg=registrasion-stripe
